requirements.txt
scikit-learn==0.19.1
tensorflow==1.8.0
certifi==2017.7.27.1
chardet==3.0.4
configparser==3.5.0
dateparser==0.6.0
docopt==0.6.2
elasticsearch==6.2.0
idna==2.6
iso8601==0.1.12
keyring==10.1
keyrings.alt==1.3
np==1.0.2
numpy==1.13.3
pyasn1==0.1.9
python-dateutil==2.6.1
pytz==2017.2
pyxdg==0.25
regex==2017.9.23
requests==2.20.0
requests-file==1.4.2
ruamel.yaml==0.15.34
SecretStorage==2.3.1
six==1.11.0
tldextract==2.1.0
tzlocal==1.4
urllib3==1.22
pandas==0.22.0
netaddr==0.7.19
lxml==4.2.1
croniter==0.3.20
millify==0.1.1
tld==0.9.1
validators==0.12.2
pygrok==1.0.0
py-spy==0.1.5


